third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/reconnect-during.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      return gattServer.getPrimaryService('heart_rate')
        .then(service => {
          let promise = assert_promise_rejects_with_message(
            service.getCharacteristic('heart_rate_measurement'),
            new DOMException(
              'GATT Server disconnected while retrieving characteristics.',
              'NetworkError'));
          gattServer.disconnect();
          return gattServer.connect().then(() => promise);
        });
    });
}, 'disconnect() and connect() called during getCharacteristic. Reject with ' +
   'NetworkError.');
</script>
